Fix debug script to exit correctly

+++ debug.sh
@@ -9,10 +9,13 @@ if [ "$1" != "build" ] && [ "$1" != "b" ]
 then
 	openocd -f interface/stlink.cfg -f target/stm32g4x.cfg -c "program ECU-25.elf" &
 	OCDPID=$!
-	#TODO Handle openocd failure correctly
-	set +e
-	arm-none-eabi-gdb --symbols=ECU-25.elf --eval-command="target extended-remote :3333"
-	kill $OCDPID
+	ps -p $OCDPID > /dev/null 2>&1
+	if [ $? -eq 0 ]
+	then
+		#set +e
+		arm-none-eabi-gdb --symbols=ECU-25.elf --eval-command="target extended-remote :3333"
+		kill $OCDPID
+	fi
 else
 	echo "Ninja exited with status ${exitcode}"
 fi
